### 2020-05-04 #39 @alegrand
[Re] Ten years challenge: Velho and Legrand (2009) - Accuracy Study and Improvement of Network Simulation in the SimGrid Framework
**Original article:** https://dl.acm.org/doi/10.4108/ICST.SIMUTOOLS2009.5592 or https://hal.inria.fr/inria-00361031/file/simutools09.pdf
```
@inproceedings{velho:inria-00361031,
  TITLE = {{Accuracy Study and Improvement of Network Simulation in the SimGrid Framework}},
  AUTHOR = {Velho, Pedro and Legrand, Arnaud},
  BOOKTITLE = {{SIMUTools'09, 2nd International Conference on Simulation Tools and Techniques}},
  ADDRESS = {Rome, Italy},
  MONTH = Mar,
  PDF = {https://hal.inria.fr/inria-00361031/file/simutools09.pdf},
  HAL_ID = {inria-00361031},
  HAL_VERSION = {v1},
}
```

**Scientific domain:**  Computer Science
**Programming language:**  C/C++
**Suggested editor:**  @khinsen or @rougier
