### 2021-03-08 #39 @benoit-girard
@rgrunbla A gentle reminder: do you consider the answer or @alegrand satisfying? Do you have additional requirements on this submission?
