### 2021-03-19 #39 @benoit-girard
@alegrand a gentle reminder that I need your updated article.pdf to proceed with publication.
