### 2021-07-23 #39 @benoit-girard
@alegrand please finalize the pdf update (https://github.com/alegrand/reproducibility-challenge/pull/1)
