### 2021-11-23 #39 @benoit-girard
@alegrand a gentle reminder that without the updated pdf your paper cannot be published.
