### 2021-12-06 #39 @alegrand
I have finally taken the time to polish this article. I deeply apologize to the editors and the reviewer (who all worked in a timely manner) for their patience.
I have just uploaded an updated version here: https://github.com/alegrand/reproducibility-challenge
The main changes are related to recent shutdown of Inria gforge. I have now ensured everything was on a perennial archive (Software Heritage and Zenodo) and I have updated the recipes. We may now proceed to the publication.
