### 2021-12-07 #39 @benoit-girard
Some information has been lost since March: I was requiring the integration of my pull request (https://github.com/alegrand/reproducibility-challenge/pull/1), that contains the necessary metadata concerning editor and reviewer identity, as well as dates of submission, acceptance and publication.
Can you do that (and remove the "under review label"), so that I can use that pdf for the end of the publication process?
