### 2022-05-24 #39 @rougier
@benoit-girard @alegrand Is the article ready to be published then? I can help if necessary.
