### 2022-05-31 #39 @alegrand
Well yes, I think so. As I said, the final version is ready but I'm not sure how to proceed with the publication.
