### 2022-05-31 #39 @rougier
@benoit-girard will proceed with the publication (or I can do it just tell me Benoît)
