### 2023-12-04 #39 @benoit-girard
OK, this got lost. I absolutely confirm we should have finalized the publication operations more than one year ago! I am really sorry, @alegrand
@rougier can you help by doing the last publication steps magic?
