### 2023-12-04 #39 @rougier
Sure, where are the latex sources?
