### 2023-12-05 #39 @alegrand
Next to the pdf I mentioned earlier. :smile:  https://github.com/alegrand/reproducibility-challenge/blob/master/article.tex :pray:
