### 2023-12-05 #39 @rougier
Ok, here is the sandbox version: https://sandbox.zenodo.org/record/6265
@alegrand @benoit-girard Can you have a look if everything's correct? If yes, I'll publish the final version.
