### 2023-12-05 #39 @benoit-girard
It's a matter of formatting, but page 2 begins with a comma, this is quite strange... Otherwise, it seems to be OK.
