### 2023-12-06 #39 @rougier
It's online !!! https://zenodo.org/record/10275726. It will soon appear on journal website.
